Wrapper function for teal modules to enable their use with dv.manager

Wrapper function for teal modules to enable their use with dv.manager

Usage

mod_teal(module_id, teal_module, j_keys = teal.data::join_keys())

Arguments

module_id

[character(1)]

Unique identifier for the module.

teal_module

[teal_module]

Teal module like teal.modules.clinical::tm_t_events()

j_keys

[join_keys]

object with dataset column relationships used for joining. Will be created by teal.data::join_keys()

Value

A list containing the following elements to be used by dv.manager:

ui: A UI function of the dv.teal module.

server: A server function of the dv.teal module.

module_id: A unique identifier.

Examples

if (FALSE) { # \dontrun{
dm <- pharmaversesdtm::dm

ae <- pharmaversesdtm::ae

data_list <- list(
  dm = dm,
  ae = ae
)

j_keys <- teal.data::join_keys(
  teal.data::join_key(
    "dm",
    keys = c("STUDYID", "USUBJID")
  ),
  teal.data::join_key(
    "ae",
    keys = c("STUDYID", "USUBJID")
  ),
  teal.data::join_key(
    "dm",
    "ae",
    keys = c("STUDYID", "USUBJID")
  )
)

mod_events <- teal.modules.clinical::tm_t_events(
  label = "Adverse Event Table",
  dataname = "ae",
  parentname = "dm",
  arm_var = choices_selected(c("ARM", "ARMCD"), "ARM"),
  llt = choices_selected(
    choices = variable_choices(ae, c("AETERM", "AEDECOD")),
    selected = c("AEDECOD")
  ),
  hlt = choices_selected(
    choices = variable_choices(ae, c("AEBODSYS", "AESOC")),
    selected = "AEBODSYS"
  ),
  add_total = TRUE,
  event_type = "adverse event"
)

module_list <- list(
  "Events" = mod_teal(
    module_id = "events",
    teal_module = mod_events,
    j_keys = j_keys
  )
)

dv.manager::run_app(
  data = list("demo" = data_list),
  module_list = module_list,
  filter_data = "dm"
)
} # }